+++ packages-next/keystone/src/lib/createAdminUIServer.ts
@@ -6,7 +6,7 @@ import express from 'express';
 import { ApolloServer } from 'apollo-server-express';
 // @ts-ignore
 import { formatError } from '@keystonejs/keystone/lib/Keystone/format-error';
-import type { KeystoneSystem } from '@keystone-next/types';
+import type { KeystoneSystem, KeystoneConfig } from '@keystone-next/types';
 
 const dev = process.env.NODE_ENV !== 'production';
 
@@ -34,15 +34,23 @@ const addApolloServer = ({ server, system }: { server: any; system: KeystoneSyst
     // FIXME: Support for generic custom apollo configuration
     // ...apolloConfig,
   });
-  // FIXME: Support for custom apiPath (config.graphql.path ?), is '/admin/api' in core keystone
-  apolloServer.applyMiddleware({ app: server, path: '/api/graphql' });
+  // FIXME: Support custom API path via config.graphql.path.
+  // Note: Core keystone uses '/admin/api' as the default.
+  apolloServer.applyMiddleware({ app: server, path: '/api/graphql', cors: false });
 };
 
-export const createAdminUIServer = async (system: KeystoneSystem) => {
+export const createAdminUIServer = async (config: KeystoneConfig, system: KeystoneSystem) => {
   const server = express();
 
-  // TODO: allow cors to be configured
-  server.use(cors({ origin: true, credentials: true }));
+  if (config.server?.cors) {
+    // Setting config.server.cors = true will provide backwards compatible defaults
+    // Otherwise, the user can provide their own config object to use
+    const corsConfig =
+      typeof config.server.cors === 'boolean'
+        ? { origin: true, credentials: true }
+        : config.server.cors;
+    server.use(cors(corsConfig));
+  }
 
   console.log('✨ Preparing Next.js app');
   const app = next({ dev, dir: Path.join(process.cwd(), '.keystone', 'admin') });