+++ tests/api-tests/fields/filter.test.js
@@ -3,7 +3,7 @@ const path = require('path');
 const { multiAdapterRunners, setupServer } = require('@keystonejs/test-utils');
 import { createItem, getItems } from '@keystonejs/server-side-graphql-client';
 
-const testModules = globby.sync(`packages/**/src/**/test-fixtures.js`, {
+const testModules = globby.sync(`{packages,packages-next}/**/src/**/test-fixtures.js`, {
   absolute: true,
 });
 testModules.push(path.resolve('packages/fields/tests/test-fixtures.js'));